@@ -5,12 +5,12 @@
     "keywords": ["enum", "phpstan"],
     "license": "BSD-3-Clause",
     "require": {
-        "php": "^7.1 | ^8.0",
-        "marc-mabe/php-enum": "^1.1 || ^2.0 || ^3.0 || ^4.0",
-        "phpstan/phpstan": "^1.0"
+        "php": "^7.4 | ^8.0",
+        "marc-mabe/php-enum": "^4.0",
+        "phpstan/phpstan": "^2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.5 | ^8.5 | 9.4"
+        "phpunit/phpunit": "^9.0"
     },
     "autoload": {
         "psr-4": {
