@@ -9,10 +9,7 @@ use PHPStan\Testing\PHPStanTestCase;
 
 class EnumDynamicReturnTypeExtensionTest extends PHPStanTestCase
 {
-    /**
-     * @var EnumDynamicReturnTypeExtension
-     */
-    protected $extension;
+    protected EnumDynamicReturnTypeExtension $extension;
 
     public function setUp(): void
     {
@@ -72,13 +69,7 @@ class EnumDynamicReturnTypeExtensionTest extends PHPStanTestCase
 
     public function staticMethodsProvider(): array
     {
-        $staticMethods = [];
-
-        if (method_exists(Enum::class, 'getValues')) {
-            $staticMethods[] = ['getValues'];
-        }
-
-        return $staticMethods;
+        return [['getValues']];
     }
 
     public function objectMethodsProvider(): array
