@@ -7,20 +7,15 @@ use MabeEnum\PHPStan\tests\assets\DeprecatedEnum;
 use MabeEnum\PHPStan\tests\assets\DocCommentEnum;
 use MabeEnum\PHPStan\tests\assets\VisibilityEnum;
 use PHPStan\Reflection\ParametersAcceptorSelector;
+use PHPStan\Reflection\ReflectionProvider;
 use PHPStan\Testing\PHPStanTestCase;
 use PHPStan\Type\VerbosityLevel;
+use PHPStan\Analyser\Scope;
 
 class EnumMethodReflectionTest extends PHPStanTestCase
 {
-    /**
-     * @var \PHPStan\Reflection\ReflectionProvider
-     */
-    protected $reflectionProvider;
-
-    /**
-     * @var EnumMethodsClassReflectionExtension
-     */
-    protected $reflectionExtension;
+    protected ReflectionProvider $reflectionProvider;
+    protected EnumMethodsClassReflectionExtension $reflectionExtension;
 
     public function setUp(): void
     {
@@ -30,7 +25,7 @@ class EnumMethodReflectionTest extends PHPStanTestCase
 
     public function testGetName(): void
     {
-        $classReflection  = $this->reflectionProvider->getClass(VisibilityEnum::class);
+        $classReflection = $this->reflectionProvider->getClass(VisibilityEnum::class);
         $methodReflection = $this->reflectionExtension->getMethod($classReflection, 'STR');
 
         $this->assertSame('STR', $methodReflection->getName());
@@ -38,7 +33,7 @@ class EnumMethodReflectionTest extends PHPStanTestCase
 
     public function testGetDeclaringClass(): void
     {
-        $classReflection  = $this->reflectionProvider->getClass(VisibilityEnum::class);
+        $classReflection = $this->reflectionProvider->getClass(VisibilityEnum::class);
         $methodReflection = $this->reflectionExtension->getMethod($classReflection, 'STR');
 
         $this->assertSame($classReflection, $methodReflection->getDeclaringClass());
@@ -46,7 +41,7 @@ class EnumMethodReflectionTest extends PHPStanTestCase
 
     public function testShouldBeStatic(): void
     {
-        $classReflection  = $this->reflectionProvider->getClass(VisibilityEnum::class);
+        $classReflection = $this->reflectionProvider->getClass(VisibilityEnum::class);
         $methodReflection = $this->reflectionExtension->getMethod($classReflection, 'STR');
 
         $this->assertTrue($methodReflection->isStatic());
@@ -54,7 +49,7 @@ class EnumMethodReflectionTest extends PHPStanTestCase
 
     public function testShouldNotBePrivate(): void
     {
-        $classReflection  = $this->reflectionProvider->getClass(VisibilityEnum::class);
+        $classReflection = $this->reflectionProvider->getClass(VisibilityEnum::class);
         $methodReflection = $this->reflectionExtension->getMethod($classReflection, 'STR');
 
         $this->assertFalse($methodReflection->isPrivate());
@@ -62,7 +57,7 @@ class EnumMethodReflectionTest extends PHPStanTestCase
 
     public function testShouldBePublic(): void
     {
-        $classReflection  = $this->reflectionProvider->getClass(VisibilityEnum::class);
+        $classReflection = $this->reflectionProvider->getClass(VisibilityEnum::class);
         $methodReflection = $this->reflectionExtension->getMethod($classReflection, 'STR');
 
         $this->assertTrue($methodReflection->isPublic());
@@ -70,7 +65,7 @@ class EnumMethodReflectionTest extends PHPStanTestCase
 
     public function testGetPrototype(): void
     {
-        $classReflection  = $this->reflectionProvider->getClass(VisibilityEnum::class);
+        $classReflection = $this->reflectionProvider->getClass(VisibilityEnum::class);
         $methodReflection = $this->reflectionExtension->getMethod($classReflection, 'STR');
 
         $this->assertSame($methodReflection, $methodReflection->getPrototype());
@@ -78,11 +73,20 @@ class EnumMethodReflectionTest extends PHPStanTestCase
 
     public function testGetVariants(): void
     {
-        $classReflection  = $this->reflectionProvider->getClass(VisibilityEnum::class);
+        $classReflection = $this->reflectionProvider->getClass(VisibilityEnum::class);
         $methodReflection = $this->reflectionExtension->getMethod($classReflection, 'STR');
-        $parametersAcceptor = ParametersAcceptorSelector::selectSingle($methodReflection->getVariants());
 
-        $this->assertSame(VisibilityEnum::class, $parametersAcceptor->getReturnType()->describe(VerbosityLevel::value()));
+        $scope = $this->createMock(Scope::class);
+        $parametersAcceptor = ParametersAcceptorSelector::selectFromArgs(
+            $scope,
+            [],
+            $methodReflection->getVariants()
+        );
+
+        $this->assertSame(
+            VisibilityEnum::class,
+            $parametersAcceptor->getReturnType()->describe(VerbosityLevel::value())
+        );
     }
 
     public function testGetDocComment(): void
@@ -151,25 +155,5 @@ class EnumMethodReflectionTest extends PHPStanTestCase
         $methodReflection = $this->reflectionExtension->getMethod($classReflection, 'STR');
 
         $this->assertTrue($methodReflection->hasSideEffects()->no());
-    }
-
-    public static function assertMatchesRegularExpression(string $pattern, string $string, string $message = ''): void
-    {
-        if (method_exists(parent::class, 'assertMatchesRegularExpression')) {
-            parent::assertMatchesRegularExpression($pattern,  $string, $message);
-            return;
-        }
-
-        self::assertRegExp($pattern, $string, $message);
-    }
-
-    public static function assertDoesNotMatchRegularExpression(string $pattern, string $string, string $message = ''): void
-    {
-        if (method_exists(parent::class, 'assertDoesNotMatchRegularExpression')) {
-            parent::assertDoesNotMatchRegularExpression($pattern,  $string, $message);
-            return;
-        }
-
-        self::assertNotRegExp($pattern, $string, $message);
     }
 }
